+++ hed/scripts/add_hed_ids.py
@@ -1,7 +1,7 @@
 from hed.scripts.script_util import get_prerelease_path
 from hed.scripts.convert_and_update_schema import convert_and_update
 import argparse
-from hed.schema.schema_io.df2schema import SchemaLoaderDF
+from hed.schema.schema_io.ontology_util import convert_filenames_to_dict
 
 # Slightly tweaked version of convert_and_update_schema.py with a new main function to allow different parameters.
 def main():
@@ -13,7 +13,7 @@ def main():
     args = parser.parse_args()
 
     basepath = get_prerelease_path(args.repo_path, schema_name=args.schema_name, schema_version=args.schema_version)
-    filenames = list(SchemaLoaderDF.convert_filenames_to_dict(basepath).values())
+    filenames = list(convert_filenames_to_dict(basepath).values())
     set_ids = True
 
     return convert_and_update(filenames, set_ids)
