+++ hed/scripts/convert_and_update_schema.py
@@ -1,5 +1,5 @@
 from hed.scripts.script_util import sort_base_schemas, validate_all_schemas, add_extension
-from hed.schema.schema_io.df2schema import load_dataframes
+from hed.schema.schema_io import load_dataframes
 from hed.schema.schema_io.ontology_util import update_dataframes_from_schema, save_dataframes
 from hed.schema.hed_schema_io import load_schema, from_dataframes
 from hed.errors import get_printable_issue_string, HedFileError
